# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatBody": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "thread_id": {
            "default": "default",
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": ["message"],
        "title": "ChatBody",
        "type": "object"
      },
      "ChatOut": {
        "properties": {
          "final_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Response"
          },
          "pending": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": ["type", "thread_id"],
        "title": "ChatOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LabelIn": {
        "properties": {
          "critique": {
            "title": "Critique",
            "type": "string"
          },
          "role": {
            "default": "adjudicator",
            "title": "Role",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "verdict": {
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": ["trace_id", "verdict", "critique"],
        "title": "LabelIn",
        "type": "object"
      },
...
      "LabelOut": {
        "properties": {
          "progress": {
            "$ref": "#/components/schemas/ProgressOut"
          }
        },
        "required": ["progress"],
        "title": "LabelOut",
        "type": "object"
      },
...
      "PendingOut": {
        "properties": {
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": ["tool", "args"],
        "title": "PendingOut",
        "type": "object"
      },
      "ProgressOut": {
        "properties": {
          "labeled": {
            "title": "Labeled",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["labeled", "total"],
        "title": "ProgressOut",
        "type": "object"
      },
...
      "ResultOut": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          }
        },
        "required": ["reference", "applied"],
        "title": "ResultOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Atlas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat": {
      "post": {
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat"
      }
    },
...
    "/labels": {
      "post": {
        "operationId": "post_label_labels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Label"
      }
    },
...
  }
}